+++ BookManagerWithSwiftUI/API/Entity/AuthEntity.swift
